Finally set correct versioning
Add-ons were failing versioning check in Kodi not because of having a hyphen, but because of having **two** hyphens. Now we change second hyphen to a tilde as the current implementation is only getting parsed in Kodi as a version (due to no hyphen at all) instead of version-revision, which is causing issues with sorting what is the latest version. I hope this is the last of this!